UPSTREAM: security/vboot: Removed vboot_prepare from vboot_locator
When prog_locate() is called in the stage VBOOT is starting from and the
image to be loaded is not the target image vboot_prepare() may be called
too early.

To prevent this vboot_prepare() is removed from the vboot_locator
structure. This allows more control over the start of the vboot logic.

To clarify the change the vboot_prepare() has been renamed to
vboot_run_logic() and calls to initialize vboot have been added at the
following places:

postcar_loader: when VBOOT starts in ROMSTAGE
romstage_loader: when VBOOT starts in BOOTBLOCK
ramstage_loader: when VBOOT starts in ROMSTAGE

BUG=N/A
TEST=tested on facebook fbg1701

Change-Id: I52cae03c94e8fa706c7eacf93f26c53120350767
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1058dd84f06fa2fcbdd99eb99da07dccdf5b9722
Original-Change-Id: Id5e8fd78458c09dd3896bfd142bd49c2c3d686df
Original-Signed-off-by: Wim Vervoorn <wvervoorn@eltan.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36543
Original-Reviewed-by: Frans Hendriks <fhendriks@eltan.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1899951
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Wim Vervoorn <wvervoorn@eltan.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Wim Vervoorn <wvervoorn@eltan.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>