UPSTREAM: security/vboot: Add rw_region_only support to vboot
In some case where the flash space is limited or when a large payload such as LinuxBoot
is used, the RO region may not be large enough to contain all components that would
normally be added.

This patch adds the possibility to add specific components to the RW regions only in
the same way as the RO_ONLY_SUPPORT does for the RO region.

Please note: this applies only to the items that would normally be added to all regions.
If the payload is directed to the RW region only, a recovery payload needs to be added
to the RO region manually.

BUG=N/A
TEST=build

Change-Id: I6cb658cf311fe69331bf9a4a8526772e8240b3df
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a1c259beef553c072b890951a328c85153035437
Original-Change-Id: Ie0df9b5dfc6df4f24efc5582a1aec9ecfb48c44d
Original-Signed-off-by: Wim Vervoorn <wvervoorn@eltan.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36544
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Frans Hendriks <fhendriks@eltan.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1910038
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Wim Vervoorn <wvervoorn@eltan.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Wim Vervoorn <wvervoorn@eltan.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>