UPSTREAM: lib/cbfs: Add fallback to RO region to cbfs_boot_locate
With this change cbfs_boot_locate will check the RO (COREBOOT) region if
a file can not be found in the active RW region. By doing so it is not
required to duplicate static files that are not intended to be updated
to the RW regions.

The coreboot image can still be updated by adding the file to the RW
region.

This change is intended to support VBOOT on systems with a small flash
device.

BUG=N/A
TEST=tested on facebook fbg1701

Change-Id: I05b5ebb71c522203a4a3cc70cc69616ceeff0031
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 114e2e88305a6e1fc972a58a03b89a23685e5a48
Original-Change-Id: I81ceaf927280cef9a3f09621c796c451e9115211
Original-Signed-off-by: Wim Vervoorn <wvervoorn@eltan.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36545
Original-Reviewed-by: Frans Hendriks <fhendriks@eltan.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1903988
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Wim Vervoorn <wvervoorn@eltan.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Wim Vervoorn <wvervoorn@eltan.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>