UPSTREAM: nb/intel/sandybridge: Make the mainboard_rcba_config hook optional
This also changes the name to mainboard_late_rcba_config to better
reflect what it does.

This adds an empty weakly linked default. The rationale behind this
change is that without an implementation of the hook some features
might not work but that the result is likely still able to boot, so it
can be made optional.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2ba0731f09d8254de0a67689e63d2d7bb179092b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9c538348d8ccaef2c3dd6b898a1f44b00ea59690
Original-Change-Id: I1897d0f5ca7427d304a425f5256cd43c088ff936
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36781
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1921917
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>