UPSTREAM: nb/intel/sandybridge: Move to C_ENVIRONMENT_BOOTBLOCK
There is some overlap between romstage and bootblock.
LPC setup and BAR initialization is now done twice.
The rationale is that the romstage should not depend too
much on the bootblock, since it can reside in a RO fmap
region.

Enabling the console will be done in a followup patch.

BUG=none
BRANCH=none
TEST=none

Change-Id: I63a24ac166a5fd801f404a9826b0514c7eaaa3e2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 360d94745feea766de7ef19487ba9158221faca0
Original-Change-Id: I4d0ba29111a5df6f19033f5ce95adcc0d9adc1fd
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36783
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1921920
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>