UPSTREAM: mb/lenovo/x60: Add vboot support
It's relatively slow to boot. It takes 1.5s to get to the payload.
In timestamps there are entries related to TPM, which are somewhat
weird given that the TPM is not enabled on this device (buggy).

TESTED: boot X60, with CONFIG_H8_FN_KEY_AS_VBOOT_RECOVERY_SW=y you
can force the recovery bootpath.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9163fde3a95544e2b9d7d6a4624686615535fb70
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b894ad5233c043f7b9c10ca3bff5cd9841c7ad64
Original-Change-Id: Ia9666194e98b7d23b97eaff08e6177684e35eca7
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/37148
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2151594
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>