UPSTREAM: mb/asrock/h110m: disable CLKREQ to use onboard LAN
The PCH uses the SRCCLKREQ# pin to detect PCIe device in the slot in
order to send clock signal to it. However, this logic is not required
for the Realtek LAN device, since this chip is soldered to the board
and always uses clocking. The chipset can't receive the clock request
signal (most likely this pin isn't connected) and doesn't enable the
CLK. For this reason, the device is broken during the initialization
phase. The patch disables clock request logic for the PCH PCIe port 6
to initialize the onboard LAN device correctly.

BUG=none
BRANCH=none
TEST=none

Change-Id: If4011da1d52c2aeb95c2c52ae11376555cf4c297
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2f50d7cd3bac409390b3e3131650a5c5b26a4e0b
Original-Change-Id: I5cbce6177c89052eb50959f43903b6f8a607e77f
Original-Signed-off-by: Maxim Polyakov <max.senia.poliak@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/36377
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1954367
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>