UPSTREAM: src: Conditionally include TEVT
ACPI method TEVT is reported as unused by iASL (20190509) when ChromeEC support is not
enabled. The message is “Method Argument is never used (Arg0)” on Method (TEVT, 1, NotSerialized),
which indicates the TEVT method is empty.

The solution is to only enable the TEVT code in mainboard or SoC when an EC is used that uses
this event. The TEVT code in the EC is only enabled if the mainboard or SoC code implements TEVT.

The TEVT method will be removed from the ASL code when the EC does not support TEVT.

BUG=N/A
TEST=Tested on facebook monolith.

Change-Id: I15332accdfdb27ecc03afea209b602b53bf9ebee
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9cb88a70f7a636806752542216e177ba625e77d2
Original-Change-Id: I8d2e14407ae2338e58797cdc7eb7d0cadf3cc26e
Original-Signed-off-by: Wim Vervoorn <wvervoorn@eltan.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/37560
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Frans Hendriks <fhendriks@eltan.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1970626
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>