UPSTREAM: mb/google/peppy: add _DSD to touchscreen ACPI
Recent changes to the Atmel touchscreen driver in the mainline
kernel broke functionality with devices running upstream coreboot,
due relying on another driver (chromeos_laptop) which makes the
assumption that the i2c devices are be in PCI mode (as with the
stock Google firmware) rather than in ACPI mode as they are in
upstream coreboot.

Mitigate this by adding the required devicetree property so the
Atmel toushcreen driver will correctly attach without the use
of chromeos_laptop.

Test: build/boot peppy on 4.18+ kernel, verify touchscreen working

BUG=none
BRANCH=none
TEST=none

Change-Id: I98f3444fcaae8df5324e9c8b69f509fe6bcf3217
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 66bd887005a5836c34228a763ce19893e644cae3
Original-Change-Id: I05df8367886eef55b409590f75a68d98d4e5fbdf
Original-Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/37915
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nicolò
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1985059
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>