UPSTREAM: soc/amd/common/block/spi: remove code duplication
This removes all the duplicated code and logic and leverages
the existing ones in libraries themselves. The current side
effect is that protection cannot be fully enabled because the
read, write, and write enable command are not exposed in struct
spi_flash currently. That support can be revised if protection
scheme makes sense for our use-cases once it's better understood.

BUG=b:146928174
Signed-off-by: Aaron Durbin <adurbin@chromium.org>

Change-Id: I21a2783022076969b4bad5de60099bf8c2c77003
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 173620a88dfe3afec1733d7bec8846406f9c8946
Original-Change-Id: I8faf9cc719ee33dd9f03fb74b579b02bbc6a5e2e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/37957
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1997416
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>